swaps related traits to swaps.rs module
