+++ libs/mocks/src/token_swaps.rs
@@ -1,6 +1,6 @@
 #[frame_support::pallet]
 pub mod pallet {
-	use cfg_traits::{OrderInfo, OrderRatio, TokenSwaps};
+	use cfg_traits::swaps::{OrderInfo, OrderRatio, TokenSwaps};
 	use frame_support::pallet_prelude::*;
 	use mock_builder::{execute_call, register_call};
