+++ pallets/order-book/src/mock.rs
@@ -11,7 +11,7 @@
 // GNU General Public License for more details.
 
 use cfg_mocks::pallet_mock_fees;
-use cfg_traits::{ConversionToAssetBalance, SwapState};
+use cfg_traits::{swaps::SwapState, ConversionToAssetBalance};
 use cfg_types::tokens::{CurrencyId, CustomMetadata};
 use frame_support::{
 	parameter_types,
