+++ pallets/order-book/src/tests.rs
@@ -10,7 +10,7 @@
 // MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
 // GNU General Public License for more details.
 
-use cfg_traits::{OrderInfo, OrderRatio, Swap, SwapState, TokenSwaps};
+use cfg_traits::swaps::{OrderInfo, OrderRatio, Swap, SwapState, TokenSwaps};
 use frame_support::{
 	assert_err, assert_noop, assert_ok,
 	traits::fungibles::{Inspect, InspectHold},
